+++ tests/deprecated_api/test_remove_1-8.py
@@ -451,6 +451,37 @@ def test_v1_8_0_remove_on_epoch_start_end_lightning_module(tmpdir):
         trainer.fit(model)
 
 
+def test_v1_8_0_remove_on_pretrain_routine_start_end_lightning_module(tmpdir):
+    class CustomModel(BoringModel):
+        def on_pretrain_routine_start(self, *args, **kwargs):
+            print("foo")
+
+    model = CustomModel()
+    trainer = Trainer(
+        fast_dev_run=True,
+        default_root_dir=tmpdir,
+    )
+    with pytest.deprecated_call(
+        match="The `LightningModule.on_pretrain_routine_start` hook was deprecated in v1.6 and will be removed in v1.8"
+    ):
+        trainer.fit(model)
+
+    class CustomModel(BoringModel):
+        def on_pretrain_routine_end(self, *args, **kwargs):
+            print("foo")
+
+    trainer = Trainer(
+        fast_dev_run=True,
+        default_root_dir=tmpdir,
+    )
+
+    model = CustomModel()
+    with pytest.deprecated_call(
+        match="The `LightningModule.on_pretrain_routine_end` hook was deprecated in v1.6 and will be removed in v1.8"
+    ):
+        trainer.fit(model)
+
+
 def test_v1_8_0_rank_zero_imports():
 
     import warnings